feat(conf): add ledmapping

+++ ESP32/cube-animations.hpp
@@ -4,8 +4,9 @@
 
 #include <Adafruit_NeoPixel.h>
 #include "rubiks-cube.hpp"
+#include "sys_conf.hpp"
 
-void chargeAnimation(Adafruit_NeoPixel* pixels, Cube* cube, Cube* ledMapping, uint8_t color, uint8_t delayval);
+void chargeAnimation(Adafruit_NeoPixel* pixels, Cube* cube, Cube* ledMapping, uint8_t color, int delayval);
 void easterAnimation(Adafruit_NeoPixel* pixels, Cube* cube, Cube* ledMapping);
 
 #endif
